[SPARK-44333][CONNECT][SQL] Move EnhancedLogicalPlan out of ParserUtils
### What changes were proposed in this pull request?
This PR moves EnhancedLogicalPlan out of ParserUtils. This will allow us to move ParserUtils to sql/api.

### Why are the changes needed?
We want to move (data type) parsing to sql/api.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
Existing tests.

Closes #41890 from hvanhovell/SPARK-44333.

Authored-by: Herman van Hovell <herman@databricks.com>
